+++ package.json
@@ -28,6 +28,7 @@
     "@smithy/protocol-http": "^3.0.12",
     "@smithy/signature-v4": "^2.0.19",
     "@tauri-apps/api": "1.5.3",
+    "@types/markdown-it": "^14.1.1",
     "@xenova/transformers": "^2.17.1",
     "blueimp-md5": "^2.19.0",
     "body-parser": "^1.20.2",
@@ -50,6 +51,7 @@
     "localforage": "^1.10.0",
     "lodash": "^4.17.21",
     "lucide-svelte": "^0.292.0",
+    "markdown-it": "^14.1.0",
     "marked": "^5.1.2",
     "ml-distance": "^4.0.1",
     "mobile-drag-drop": "3.0.0-rc.0",
