+++ src/ts/translator/translator.ts
@@ -5,10 +5,10 @@ import { globalFetch, isTauri } from "../storage/globalApi"
 import { alertError } from "../alert"
 import { requestChatData } from "../process/request"
 import { doingChat } from "../process"
-import type { simpleCharacterArgument } from "../parser"
+import { applyMarkdownToNode, type simpleCharacterArgument } from "../parser"
 import { selectedCharID } from "../stores"
 import { getModuleRegexScripts } from "../process/modules"
-import { getNodetextToSentence, sleep, applyMarkdownToNode } from "../util"
+import { getNodetextToSentence, sleep } from "../util"
 import { processScriptFull } from "../process/scripts"
 import { Capacitor } from "@capacitor/core"
